feat: users can visit player profile from match detail resolves #138
