+++ lib/screens/player_detail/player_detail_header.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:paladinsedge/models/index.dart' as models;
 import 'package:paladinsedge/widgets/index.dart' as widgets;
 
-class PlayerDetailComponent extends StatelessWidget {
+class PlayerDetailHeader extends StatelessWidget {
   final models.Player player;
   final void Function() onForceUpdate;
   final bool isLoading;
 
-  const PlayerDetailComponent({
+  const PlayerDetailHeader({
     required this.player,
     required this.onForceUpdate,
     required this.isLoading,
